+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/legacy_property_info.php
@@ -2,6 +2,6 @@
 
 $container->loadFromExtension('framework', [
     'property_info' => [
-        'enabled' => true,
+        'with_constructor_extractor' => false,
     ],
 ]);