+++ src/Symfony/Component/PropertyInfo/Tests/DependencyInjection/Fixtures/property_info_disabled.php
@@ -1,0 +1,3 @@
+<?php
+
+$container->loadFromExtension('property_info', ['enabled' => false]);